@@ -2,31 +2,29 @@ apply plugin: 'com.android.library'
 
 apply plugin: 'kotlin-android'
 
-apply plugin: 'kotlin-android-extensions'
-
-apply from: "../ktlint.gradle"
+apply plugin: "guru.stefma.bintrayrelease"
 
 android {
-    compileSdkVersion 29
-
+    compileSdkVersion 30
 
     defaultConfig {
         minSdkVersion 19
-        targetSdkVersion 29
-        versionCode 9
-        versionName "1.7.1"
+        targetSdkVersion 30
+        versionCode 7
+        versionName "1.6"
 
         testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
     }
 
     buildTypes {
-        debug{
-
-        }
         release {
             minifyEnabled false
             proguardFiles getDefaultProguardFile('proguard-android-optimize.txt'), 'proguard-rules.pro'
         }
+    }
+
+    buildFeatures {
+        viewBinding true
     }
 
     sourceSets {
@@ -44,48 +42,26 @@ dependencies {
     implementation fileTree(dir: 'libs', include: ['*.jar'])
     implementation"org.jetbrains.kotlin:kotlin-stdlib-jdk7:$kotlin_version"
 
-    implementation 'androidx.core:core-ktx:1.2.0'
-    implementation 'androidx.appcompat:appcompat:1.1.0'
-    implementation "androidx.exifinterface:exifinterface:1.1.0"
+    implementation 'androidx.core:core-ktx:1.3.2'
+    implementation 'androidx.appcompat:appcompat:1.2.0'
+    implementation "androidx.exifinterface:exifinterface:1.3.2"
 
     //More Info: https://github.com/Yalantis/uCrop
-    implementation 'com.github.yalantis:ucrop:2.2.4'
+    implementation 'com.github.yalantis:ucrop:2.2.5'
 
     //More Info: https://github.com/florent37/InlineActivityResult
-    compileOnly 'com.github.florent37:inline-activity-result-kotlin:1.0.3'
+    compileOnly 'com.github.florent37:inline-activity-result-kotlin:1.0.2'
 
-    testImplementation 'junit:junit:4.13'
-    androidTestImplementation 'androidx.test.ext:junit:1.1.1'
-    androidTestImplementation 'androidx.test:core:1.2.0'
+    testImplementation 'junit:junit:4.12'
+    androidTestImplementation 'androidx.test.ext:junit:1.1.2'
+    androidTestImplementation 'androidx.test:core:1.3.0'
 }
 
-ext {
-    bintrayRepo = 'maven'
-    bintrayName = 'imagepicker'
-
-    publishedGroupId = 'com.github.dhaval2404'
-    libraryName = 'imagepicker'
-    artifact = 'imagepicker'
-
-    libraryDescription = 'Pick image from Gallery or Capture new image with Camera.'
-
-    siteUrl = 'https://github.com/Dhaval2404/ImagePicker/'
-    gitUrl = 'https://github.com/Dhaval2404/ImagePicker.git'
-
-    libraryVersion = '1.7.1'
-    //If you are uploading new library try : gradlew install
-    //If you are updating existing library then execute: gradlew bintrayUpload
-    //In both the case don't forgot to put bintray credentials in local.properties file.
-
-    developerId = 'dhaval2404'
-    developerName = 'Dhaval Patel'
-    developerEmail = 'dhavalpatel244@gmail.com'
-
-    licenseName = 'The Apache Software License, Version 2.0'
-    licenseUrl = 'http://www.apache.org/licenses/LICENSE-2.0.txt'
-    allLicenses = ["Apache-2.0"]
+publish {
+    version = "1.12"
+    group = "com.thejuki"
+    userOrg = 'thejuki'
+    artifactId = 'imagepicker'
+    desc = 'Fix ImagePickerActivity Opening Gallery or Camera Twice from original project.'
+    website = 'https://github.com/Dhaval2404/ImagePicker/pull/63'
 }
-
-// Place it at the end of the file
-apply from: 'https://raw.githubusercontent.com/nuuneoi/JCenter/master/installv1.gradle'
-apply from: 'https://raw.githubusercontent.com/nuuneoi/JCenter/master/bintrayv1.gradle'
